Refactor export functionality in PartialImportExport module
