+++ app/client/cypress/e2e/Regression/ClientSide/PartialImportExport/PartialExport_spec.ts
@@ -30,98 +30,52 @@ describe(
     });
 
     beforeEach(() => {
-      entityExplorer.ActionContextMenuByEntityName({
-        entityNameinLeftSidebar: "Home",
-        action: "Export",
-        entityType: EntityItems.Page,
-      });
-
-      agHelper.AssertElementVisibility(
-        partialImportExport.locators.export.exportModal,
-      );
-      agHelper.AssertElementEnabledDisabled(
-        partialImportExport.locators.export.modelContents.exportButton,
-      );
+      partialImportExport.OpenExportMenu();
     });
 
     it("1. Should export all the selected JsObjects", () => {
-      exportAndCompareDownloadedFile(
+      partialImportExport.ExportAndCompareDownloadedFile(
         0,
         partialImportExport.locators.export.modelContents.jsObjectsSection,
         "JSExportedOnly.json",
+        fixtureName,
       );
     });
 
     it("2. Should export all the selected datasources", () => {
-      exportAndCompareDownloadedFile(
+      partialImportExport.ExportAndCompareDownloadedFile(
         1,
         partialImportExport.locators.export.modelContents.datasourcesSection,
         "DatasourceExportedOnly.json",
+        fixtureName,
       );
     });
 
     it("3. Should export all the selected queries", () => {
-      exportAndCompareDownloadedFile(
+      partialImportExport.ExportAndCompareDownloadedFile(
         2,
         partialImportExport.locators.export.modelContents.queriesSection,
         "QueriesExportedOnly.json",
+        fixtureName,
       );
     });
 
     it("4. Should export all the customjs libs", () => {
-      exportAndCompareDownloadedFile(
+      partialImportExport.ExportAndCompareDownloadedFile(
         3,
         partialImportExport.locators.export.modelContents.customJSLibsSection,
         "CustomJsLibsExportedOnly.json",
+        fixtureName,
       );
     });
 
     it("5. Should export all the widgets", () => {
-      exportAndCompareDownloadedFile(
+      partialImportExport.ExportAndCompareDownloadedFile(
         4,
         partialImportExport.locators.export.modelContents.widgetsSection,
         "WidgetsExportedOnly.json",
+        fixtureName,
       );
     });
   },
 );
-
-function exportAndCompareDownloadedFile(
-  sectionIndex: number,
-  sectionSelector: string,
-  fileNameToCompareWith: string,
-) {
-  agHelper.GetNClick(
-    partialImportExport.locators.export.modelContents.sectionHeaders,
-    sectionIndex,
-  );
-
-  const currentSection = agHelper.GetElement(sectionSelector);
-
-  const checkboxesInSection = currentSection.find("input[type='checkbox']");
-  checkboxesInSection.each((element) => {
-    cy.wrap(element).click({ force: true });
-  });
-
-  agHelper.AssertElementEnabledDisabled(
-    partialImportExport.locators.export.modelContents.exportButton,
-    0,
-    false,
-  );
-  agHelper.GetNClick(
-    partialImportExport.locators.export.modelContents.exportButton,
-  );
-
-  cy.readFile(`cypress/downloads/${fixtureName}`).then((exportedFile) => {
-    cy.fixture(`PartialImportExport/${fileNameToCompareWith}`).then(
-      (expectedFile) => {
-        // sort the contents of both the files before comparing
-        exportedFile = JSON.stringify(exportedFile).split("").sort().join("");
-        expectedFile = JSON.stringify(expectedFile).split("").sort().join("");
-
-        expect(exportedFile).to.deep.equal(expectedFile);
-      },
-    );
-  });
-  cy.exec(`rm cypress/downloads/${fixtureName}`);
-}
