+++ app/client/cypress/support/Pages/PartialImportExport.ts
@@ -1,3 +1,6 @@
+import { ObjectsRegistry } from "../Objects/Registry";
+import { EntityItems } from "./AssertHelper";
+
 export default class PartialImportExport {
   public readonly locators = {
     import: {
@@ -20,4 +23,61 @@ export default class PartialImportExport {
       },
     },
   };
+
+  OpenExportMenu() {
+    ObjectsRegistry.EntityExplorer.ActionContextMenuByEntityName({
+      entityNameinLeftSidebar: "Home",
+      action: "Export",
+      entityType: EntityItems.Page,
+    });
+
+    ObjectsRegistry.AggregateHelper.AssertElementVisibility(
+      this.locators.export.exportModal,
+    );
+    ObjectsRegistry.AggregateHelper.AssertElementEnabledDisabled(
+      this.locators.export.modelContents.exportButton,
+    );
+  }
+
+  ExportAndCompareDownloadedFile(
+    sectionIndex: number,
+    sectionSelector: string,
+    fileNameToCompareWith: string,
+    fixtureName: string,
+  ) {
+    ObjectsRegistry.AggregateHelper.GetNClick(
+      this.locators.export.modelContents.sectionHeaders,
+      sectionIndex,
+    );
+
+    const currentSection =
+      ObjectsRegistry.AggregateHelper.GetElement(sectionSelector);
+
+    const checkboxesInSection = currentSection.find("input[type='checkbox']");
+    checkboxesInSection.each((element) => {
+      cy.wrap(element).click({ force: true });
+    });
+
+    ObjectsRegistry.AggregateHelper.AssertElementEnabledDisabled(
+      this.locators.export.modelContents.exportButton,
+      0,
+      false,
+    );
+    ObjectsRegistry.AggregateHelper.GetNClick(
+      this.locators.export.modelContents.exportButton,
+    );
+
+    cy.readFile(`cypress/downloads/${fixtureName}`).then((exportedFile) => {
+      cy.fixture(`PartialImportExport/${fileNameToCompareWith}`).then(
+        (expectedFile) => {
+          // sort the contents of both the files before comparing
+          exportedFile = JSON.stringify(exportedFile).split("").sort().join("");
+          expectedFile = JSON.stringify(expectedFile).split("").sort().join("");
+
+          expect(exportedFile).to.deep.equal(expectedFile);
+        },
+      );
+    });
+    cy.exec(`rm cypress/downloads/${fixtureName}`);
+  }
 }
